Re-organise structure and create new example files.
