+++ examples/ios/swift/Migration/AppDelegate.swift
@@ -34,12 +34,13 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         window?.rootViewController = UIViewController()
         window?.makeKeyAndVisible()
         
-        // The RealmFile_vx structs are used to create the .realm files which are then used by the app to showcase different migrations.
-        // This line and the corresponding file for schema version x has to be uncommented to create the file.
-        // One .realm file per schema version is already included in the project.
-//        RealmFile.create()
-        
-        MigrationExample.execute()
+        // Choose to either create example files or execute them.
+        let shouldCreateExamples = false
+        if shouldCreateExamples {
+            RealmTemplate.create()
+        } else {
+            MigrationExample.execute()
+        }
         
         return true
     }
