+++ examples/ios/swift/Migration/MigrationExample.swift
@@ -1,51 +1,33 @@
+////////////////////////////////////////////////////////////////////////////
 //
-//  Migration.swift
-//  Migration
+// Copyright 2021 Realm Inc.
 //
-//  Created by Dominic Frei on 28/01/2021.
-//  Copyright © 2021 Realm. All rights reserved.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+////////////////////////////////////////////////////////////////////////////
 
 import Foundation
 import RealmSwift
 
 struct MigrationExample {
     
-    static func execute() {        
-        // Define a migration block.
-        // You can define this inline, but we will reuse this to migrate realm files from multiple versions
-        // to the most current version of our data model.
-        let migrationBlock = createMigrationBlock()
-        
-        for realmVersion in RealmVersion.allVersions {
-            let realmUrl = realmVersion.destinationUrl(usingTemplate: true)
-            let realmConfiguration = Realm.Configuration(fileURL: realmUrl, schemaVersion: 3, migrationBlock: migrationBlock)
+    static func execute() {
+        for realmVersion in RealmVersion.allCases {
+            let realmUrl = URL(for: realmVersion, usingTemplate: true)
+            let schemaVersion = RealmVersion.mostRecentVersion
+            let realmConfiguration = Realm.Configuration(fileURL: realmUrl, schemaVersion: schemaVersion, migrationBlock: migrationBlock)
             try! Realm.performMigration(for: realmConfiguration)
         }
-    }
-    
-    static func createMigrationBlock() -> MigrationBlock {
-        let migrationBlock: MigrationBlock = { migration, oldSchemaVersion in
-            if oldSchemaVersion < 1 {
-                migration.enumerateObjects(ofType: Person.className()) { oldObject, newObject in
-                    // combine name fields into a single field
-                    let firstName = oldObject!["firstName"] as! String
-                    let lastName = oldObject!["lastName"] as! String
-                    newObject!["fullName"] = "\(firstName) \(lastName)"
-                }
-            }
-            if oldSchemaVersion < 2 {
-                migration.enumerateObjects(ofType: Person.className()) { _, newObject in
-                    // Add a pet to a specific person
-                    if newObject!["fullName"] as! String == "John Smith" {
-                        let johnsDog = migration.create(Dog.className(), value: ["Jimbo"])
-                        let dogs = newObject!["dogs"] as! List<MigrationObject>
-                        dogs.append(johnsDog)
-                    }
-                }
-            }
-        }
-        return migrationBlock
     }
     
 }
